Some errors related to the work of Hibernate have been fixed

+++ src/main/java/services/StudentService.java
@@ -13,7 +13,7 @@ public class StudentService {
     public StudentService() {
     }
 
-    public Student findStudent(long id) {
+    public Student findStudent(int id) {
         return studentDAO.findById(id);
     }
